Fix vim.g.terminal_color_15
`white` is not defined anymore in `colors.lua`, causing
`terminal_color_15` to be effectively nil.

This commit ties the color 15 (also known as ANSI bright white) to the
`base8` from the palette, which is either white-ish or dark-ish,
depending on selected background. This behavior is consistent with Doom
Emacs, as hard-coded at
https://github.com/doomemacs/themes/blob/ff26f26ea3d761375f5fc4070438fbd0f3473d33/doom-themes-base.el#L201